Update bst.py
I am new to data structures so correct me if I'm wrong! I think you have some unnecessary code in your _left_rotation and _right_rotation methods. After setting pivot.left = pivot.right, the parent is already guaranteed to be the pivot, so you don't need to update it with the if block. Right? Also, the last line of code in each method doesn't need to update the parent of the pivot because the parent does not change … I think. Unless it is for some special case maybe? I am still learning so take my suggestions with a grain of salt. Thanks!